https://github.com/metanorma/metanorma/issues/401
