+++ lib/metanorma/collection/config/config.rb
@@ -126,27 +126,6 @@ module Metanorma
           end
         end
 
-        def documents_from_xml(model, value)
-          model.documents = value
-            .each_with_object([]) do |b, m|
-            m << Bibdata.from_xml(b.to_xml)
-          end
-        end
-
-        def documents_to_xml(model, parent, doc)
-          doc.parent.elements.detect do |x|
-            x.name == "doc-container"
-          end and return
-          b = Nokogiri::XML::Builder.new do |xml|
-            xml.document do |m|
-              model.collection.doccontainer(m) or return
-            end
-          end
-          b.parent.elements.first.elements.each do |x|
-            doc.add_element(parent, x)
-          end
-        end
-
         include Converters
       end
     end
