+++ spec/fixtures/collection/bilingual.presentation.xml
@@ -113,8 +113,11 @@
     </structuredidentifier>
   </ext>
 </bibdata></entry></entry>
+<!--
 <format>html</format>
-<coverpage>cover.html</coverpage><prefatory-content><clause id="_clause" inline-header="false" obligation="normative">
+<coverpage>cover.html</coverpage>
+-->
+<prefatory-content><clause id="_clause" inline-header="false" obligation="normative">
 <title>Clause</title>
 <p id="_da34b037-83c7-1b19-ceae-35673a6ae11d">Welcome to our collection</p>
 </clause>
